Open device on change_state
g_error leads to call abort() and coredump.
And usually g_error is used when there is fatal error.
And should not be used in such way.

It is better to try to open device on change_state
and return gst error, and close pipeline, without a crash.

Added vimbacamera_set_camera_id/vimbacamera_get_camera_id
to be able to control camera id.
And props should not be accessed directly.